requirements.txt
fastapi
uvicorn
langgraph
langsmith
langchain
langchain_groq
langchain_community
arxiv
wikipedia
python-dotenv